FashionGANs/Conditionals/CGAN/requirements.txt
matplotlib==3.5.1
IPython~=7.26.0
tensorboard
transformers
h5py
torch~=1.9.0
torchvision~=0.10.0
numpy~=1.20.3
Pillow~=8.3.1
scipy